polymesh scene
Created a 3D environment using three different objects (a tea pot, and
two cubes), you can move forward and backward by pressing the K and J
keys, respectively. You can also rotate by pressing D for
counter-clockwise rotation, and F for clockwise rotation. The file that
must be written when prompted to "enter filename: " is called
"test.txt", which contains the direction for setting up the matrices and
the viewing angles, and it also contains the paths for the objects to be
drawn on to the window.